requirements.txt
Django==4.2.4
django_mptt==0.14.0
pytils==0.4.1
